Move SourceConstructorSymbolBase into its own file. Also follow up on rebase.
