+++ src/Compilers/CSharp/Portable/Symbols/Synthesized/Records/SynthesizedRecordPropertySymbol.cs
@@ -38,7 +38,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
                 GeneratedNames.MakeBackingFieldName(name),
                 isReadOnly: true,
                 isStatic: false,
-                hasInitializer: backingParameter.HasExplicitDefaultValue);
+                hasInitializer: true);
             GetMethod = new GetAccessorSymbol(this, name);
             SetMethod = new InitAccessorSymbol(this, name, diagnostics);
         }
